+++ changelog.md
@@ -2,3 +2,7 @@
 
 - Matched command rules with carpet
 - Performance improvement
+
+## Bugfix
+
+- (1.21.4+) legacyHoneyBlockSliding causing living entities to slide slower than expected